edupilot-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.10.3
supabase==2.4.3
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
